Support for latest plugin uninstaller
Adds the ability to test network activation, brining us in line with
recent changes in the plugin uninstall tester. We are able to simplify
things by removing the uninstall method. We allow WordPoints itself to
be deactivated and uninstalled, instead of just the module.

Also, if either of our `system()` calls fail, we fail the test now.

Fixes #9